VAC 패턴 적용(components/DiaryForm/) #9
